整理代码， 修改协议文档，ID_NULL=-1，判断需要加强制转换

+++ api/api_uart.c
@@ -195,7 +195,7 @@ void api_uarts_init(void)
 {
 	uint8_t id;
 	for(id=0; id<m_uart_num; id++){
-		if(UART_DEBUG_ID != id){
+		if((uint8_t)UART_DEBUG_ID != id){
 			api_uart_init(id);
 		}
 	}
